@@ -10,7 +10,6 @@
     ],
     "homepage": "https://github.com/rappasoft/laravel-livewire-tables",
     "license": "MIT",
-    "type": "library",
     "authors": [
         {
             "name": "Anthony Rappa",
@@ -19,16 +18,17 @@
         }
     ],
     "require": {
-        "php": "^7.3",
-        "livewire/livewire": "^1.3|^2.0"
+        "php": "^7.4|^8.0",
+        "livewire/livewire": "^2.0",
+        "spatie/laravel-package-tools": "^1.4.3",
+        "illuminate/contracts": "^8.0"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^2.16",
-        "orchestra/testbench": "^5.0",
-        "phpunit/phpunit": "^9.0"
-    },
-    "suggest": {
-        "maatwebsite/excel": "Required to use export functionality"
+        "ext-sqlite3": "*",
+        "orchestra/testbench": "^6.13",
+        "phpunit/phpunit": "^9.3",
+        "spatie/laravel-ray": "^1.9",
+        "vimeo/psalm": "^4.4"
     },
     "autoload": {
         "psr-4": {
@@ -41,9 +41,9 @@
         }
     },
     "scripts": {
-        "test": "vendor/bin/phpunit",
-        "test-coverage": "vendor/bin/phpunit --coverage-html coverage",
-        "format": "vendor/bin/php-cs-fixer fix --allow-risky=yes"
+        "psalm": "vendor/bin/psalm",
+        "test": "vendor/bin/phpunit --colors=always",
+        "test-coverage": "vendor/bin/phpunit --coverage-html coverage"
     },
     "config": {
         "sort-packages": true
@@ -51,8 +51,10 @@
     "extra": {
         "laravel": {
             "providers": [
-              "Rappasoft\\LaravelLivewireTables\\LaravelLivewireTablesServiceProvider"
+                "Rappasoft\\LaravelLivewireTables\\LaravelLivewireTablesServiceProvider"
             ]
         }
-    }
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
