@@ -1,15 +1,27 @@
-@if ($paginationEnabled)
-    <div class="row">
-        <div class="col">
-            {{ $models->links() }}
-        </div>
+@if ($showPagination)
+    @if ($paginationEnabled && $rows->lastPage() > 1)
+        <div class="row">
+            <div class="col-12 col-md-6">
+                {{ $rows->links() }}
+            </div>
 
-        <div class="col text-right text-muted">
-            @lang('laravel-livewire-tables::strings.results', [
-                'first' => $models->count() ? $models->firstItem() : 0,
-                'last' => $models->count() ? $models->lastItem() : 0,
-                'total' => $models->total()
-            ])
+            <div class="col-12 col-md-6 text-center text-md-right text-muted">
+                @lang('Showing')
+                <strong>{{ $rows->count() ? $rows->firstItem() : 0 }}</strong>
+                @lang('to')
+                <strong>{{ $rows->count() ? $rows->lastItem() : 0 }}</strong>
+                @lang('of')
+                <strong>{{ $rows->total() }}</strong>
+                @lang('results')
+            </div>
         </div>
-    </div>
+    @else
+        <div class="row">
+            <div class="col-12 text-muted">
+                @lang('Showing')
+                <strong>{{ $rows->count() }}</strong>
+                @lang('results')
+            </div>
+        </div>
+    @endif
 @endif
